recv/requirements.txt
pyserial==3.4
pandas==0.23.1
matplotlib==2.1.1
PyInstaller==3.3.1
